font remove to external sdcard

+++ src/com/industry/printer/MainActivity.java
@@ -69,6 +69,7 @@ import com.industry.printer.Utils.FileUtil;
 import com.industry.printer.Utils.PlatformInfo;
 import com.industry.printer.Utils.SystemPropertiesProxy;
 import com.industry.printer.Utils.ToastUtil;
+import com.industry.printer.Utils.ZipUtil;
 import com.industry.printer.hardware.ExtGpio;
 import com.industry.printer.hardware.FpgaGpioOperation;
 import com.industry.printer.hardware.PWMAudio;
@@ -589,11 +590,16 @@ public class MainActivity extends Activity implements OnCheckedChangeListener, O
 	 * import from USB to flash
 	 */
 	private void msgImportOnly() {
-		mProgressDialog = LoadingDialog.show(this, R.string.strCopying);
+		
 		final ArrayList<String> usbs = ConfigPath.getMountedUsb();
-
+		if (usbs.size() <= 0) {
+			ToastUtil.show(mContext, "Please insert USB device");
+			return;
+		}
+		mProgressDialog = LoadingDialog.show(this, R.string.strCopying);
+		
 		FileUtil.deleteFolder(Configs.QR_LAST);
-		Observable.just(Configs.SYSTEM_CONFIG_MSG_PATH, Configs.PICTURE_SUB_PATH, Configs.SYSTEM_CONFIG_DIR)
+		Observable.just(Configs.SYSTEM_CONFIG_MSG_PATH, Configs.PICTURE_SUB_PATH, Configs.SYSTEM_CONFIG_DIR, Configs.FONT_DIR)
 				.flatMap(new Func1<String, Observable<Map<String, String>>>() {
 
 					@Override
@@ -609,6 +615,9 @@ public class MainActivity extends Activity implements OnCheckedChangeListener, O
 						} else if ( Configs.SYSTEM_CONFIG_DIR.equals(arg0)) {
 							src.put("source",usbs.get(0) + arg0);
 							src.put("dest", Configs.CONFIG_PATH_FLASH + Configs.SYSTEM_CONFIG_DIR);
+						} else if (Configs.FONT_DIR.equals(arg0)) {
+							src.put("source",usbs.get(0) + File.separator + Configs.FONT_ZIP_FILE);
+							src.put("dest", Configs.CONFIG_PATH_FLASH + File.separator + Configs.FONT_ZIP_FILE);
 						}
 						Debug.d(TAG, "--->flatMap");
 						return Observable.just(src);
@@ -620,7 +629,10 @@ public class MainActivity extends Activity implements OnCheckedChangeListener, O
 					public Observable<Void> call(Map<String, String> arg0) {
 						try {
 							FileUtil.copyDirectiory(arg0.get("source"), arg0.get("dest"));
-							
+							String dest = arg0.get("dest");
+							if (dest.endsWith(Configs.FONT_ZIP_FILE)) {
+								ZipUtil.UnZipFolder(Configs.CONFIG_PATH_FLASH + File.separator + Configs.FONT_ZIP_FILE, Configs.CONFIG_PATH_FLASH);
+							}
 						} catch (Exception e) {
 							// TODO: handle exception
 						}
@@ -676,10 +688,15 @@ public class MainActivity extends Activity implements OnCheckedChangeListener, O
 	 * import from USB to flash
 	 */
 	private void msgImport() {
-		mProgressDialog = LoadingDialog.show(this, R.string.strCopying);
 		final ArrayList<String> usbs = ConfigPath.getMountedUsb();
+		if (usbs.size() <= 0) {
+			ToastUtil.show(mContext, "Please insert USB device");
+			return;
+		}
+		mProgressDialog = LoadingDialog.show(this, R.string.strCopying);
 		
-		Observable.just(Configs.SYSTEM_CONFIG_MSG_PATH, Configs.PICTURE_SUB_PATH, Configs.SYSTEM_CONFIG_DIR)
+		
+		Observable.just(Configs.SYSTEM_CONFIG_MSG_PATH, Configs.PICTURE_SUB_PATH, Configs.SYSTEM_CONFIG_DIR, Configs.FONT_DIR)
 		.flatMap(new Func1<String, Observable<Map<String, String>>>() {
 
 			@Override
@@ -695,6 +712,9 @@ public class MainActivity extends Activity implements OnCheckedChangeListener, O
 				} else if ( Configs.SYSTEM_CONFIG_DIR.equals(arg0)) {
 					src.put("source",usbs.get(0) + arg0);
 					src.put("dest", Configs.CONFIG_PATH_FLASH + Configs.SYSTEM_CONFIG_DIR);
+				} else if (Configs.FONT_DIR.equals(arg0)) {
+					src.put("source",usbs.get(0) + File.separator + Configs.FONT_ZIP_FILE);
+					src.put("dest", Configs.CONFIG_PATH_FLASH + File.separator + Configs.FONT_ZIP_FILE);
 				}
 				Debug.d(TAG, "--->flatMap");
 				return Observable.just(src);
@@ -706,6 +726,10 @@ public class MainActivity extends Activity implements OnCheckedChangeListener, O
 			public Observable<Void> call(Map<String, String> arg0) {
 				try {
 				FileUtil.copyClean(arg0.get("source"), arg0.get("dest"));
+				String dest = arg0.get("dest");
+				if (dest.endsWith(Configs.FONT_ZIP_FILE)) {
+					ZipUtil.UnZipFolder(Configs.CONFIG_PATH_FLASH + File.separator + Configs.FONT_ZIP_FILE, Configs.CONFIG_PATH_FLASH);
+				}
 				} catch (Exception e) {
 					// TODO: handle exception
 				}
@@ -764,9 +788,13 @@ public class MainActivity extends Activity implements OnCheckedChangeListener, O
 	 * export out to USB from flash
 	 */
 	private void msgExport() {
-		mProgressDialog = LoadingDialog.show(this, R.string.strCopying);
 		final ArrayList<String> usbs = ConfigPath.getMountedUsb();
+		if (usbs.size() <= 0) {
+			ToastUtil.show(mContext, "Please insert USB device");
+			return;
+		}
 		
+		mProgressDialog = LoadingDialog.show(this, R.string.strCopying);
 		Observable.just(Configs.SYSTEM_CONFIG_MSG_PATH, Configs.PICTURE_SUB_PATH, Configs.SYSTEM_CONFIG_DIR, "print.bin")
 		.flatMap(new Func1<String, Observable<Map<String, String>>>() {
